remote all the warning

+++ src/kernel.c
@@ -17,7 +17,7 @@
 static uint32_t user_stack[1024];
 void user_func(void) {
     const char *str = "Hello, from Ring3\n";
-    sys_call(SYS_WRITE, str, 0, 0);
+    sys_call(SYS_WRITE, (uint32_t)str, 0, 0);
 
     while (1);
 }
